# docs/swagger.yaml
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Soand API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Soand API
  version: "1.0"
paths:
...
  /chat/direct:
    get:
      consumes:
      - application/json
      description: Establishes a WebSocket connection for real-time messaging between
        two users
      parameters:
      - description: Recipient's user ID
        in: query
        name: recipient_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Missing or invalid recipient ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: WebSocket upgrade failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: WebSocket connection for real-time chat
      tags:
      - chat
  /chat/direct/messages:
    get:
      consumes:
      - application/json
      description: Retrieves paginated messages between the authenticated user and
        another user
      parameters:
      - description: Recipient's user ID
        in: query
        name: recipient_id
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: string
      - description: 'Number of messages per page (default: 10)'
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns paginated messages
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid recipient ID or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not fetch messages
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get chat messages
      tags:
      - chat
  /chat/dlete:
    delete:
      consumes:
      - application/json
      description: Deletes a specific message for the authenticated user
      parameters:
      - description: Message ID to delete
        in: path
        name: message_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid message ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not authorized to delete this message
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Message not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not delete message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a message
      tags:
      - chat
  /chat/update:
    patch:
      consumes:
      - application/json
      description: Updates the content of a specific message for the authenticated
        user
      parameters:
      - description: Message ID to update
        in: path
        name: message_id
        required: true
        type: string
      - description: New message text
        in: body
        name: message
        required: true
        schema:
          properties:
            new_text:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Message updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid message ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not authorized to update this message
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Message not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not update message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a message
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    description: '"JWT Authorization header using the Bearer scheme. Example: ''Bearer
      {token}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"